Make secret-number choices unbounded, per CR 107.1a/b
Addresses the three blockers on the previous head.

1. "a number 0 or greater" was capped at 20

The parser lowered every bare number choice to NumberRange { min: 0, max: 20 },
so 21 was rejected outright. On Wheel of Misfortune the magnitude of the number
IS the decision, so an invented ceiling made a legal choice illegal.

NumberRange's max is now Option<u32>, and None means what the rules mean: no
maximum. Bounded card text ("a number between 1 and 5") keeps Some and
serializes byte-identically -- only the unbounded form omits the key. The
between-form now DECLINES rather than substituting a ceiling when the upper
token is missing, so a malformed phrase strict-fails instead of inventing a
bound.

The stored value widens u8 -> u32 across ChosenAttribute::Number /
RevealedNumber and ChoiceValue::Number; u8 could not hold 256, let alone a real
bid. The accepted domain is bounded at i32::MAX, which is not a UI cap but the
engine's own arithmetic domain -- every quantity resolves through i32, and
damage and life totals are i32, so a number beyond it could not be dealt or
compared. Within that domain every value the rules permit is accepted.

End to end: an unbounded range enumerates nothing (compute_options returns
empty) and routes through options_supplied_by_player -- the same free-entry path
CardName already used. ChoiceType::accepts_free_entry_answer is the single
validation authority, shared by the ChooseOption answer seam and the AI's
legal-action enumeration so the two cannot disagree about what is legal. The AI
samples a life-total-anchored ladder for a domain it cannot enumerate, filtered
through that same authority. The client renders a numeric input instead of a
button grid when max is absent.

New regression: a_number_past_the_old_ceiling_is_choosable_and_deals_that_much_damage
bids 40 and 21 -- both past the old ceiling -- and asserts 40 is accepted,
stored, folded as the cross-player maximum, and dealt as 40 damage. The existing
three-seat test structurally could not detect this: it only ever chooses 1 and 4.

2. Persistence missed chosen-number reads in conditions

definition_reads_player_chosen_number walked player_scope, effect quantities,
DamageEachPlayer and sub/else links but never AbilityDefinition::condition, so a
QuantityCheck reading PlayerChosenNumber left the upstream choice
non-persistent and the answer was cleared before the condition resolved. Added a
recursive walker through QuantityCheck / And / Or / Not / ConditionInstead, plus
a regression that buries the reference under Not(And(...)) and carries a control
proving it measures the condition walk rather than blanket promotion.

3. Reveal grammar was prefix-accepting and active-voice-blind

parse_reveal_chosen_numbers_clause matched only tag("reveal ") and accepted a
prefix, discarding any remainder. It now requires complete-clause consumption
and covers third person with the "s" as its own opt axis rather than duplicated
tags. The anti-swallow test asserts the surviving tail rather than the head's
shape: the clause splitter separates a trailing instruction before this grammar
sees it, so asserting "the head is not a reveal" would test the splitter's
boundary choice instead of the property that matters.

Two things caught by the repo's own instruments while doing this:

- The extended serde round-trip test caught that an unbounded range serialized
  as "max": null rather than omitting the key. ChoiceType has a hand-written
  Serialize, so the skip_serializing_if on the ChoiceTypeData deserialize mirror
  never applied to the write path; it is now mirrored by hand.
- The committed-guess placeholder sentinel was min: 0, max: 0, which would have
  become min: 0, max: None -- the exact shape of a genuine unbounded choice,
  classifying every real one as an unfilled placeholder. The sentinel is now
  Some(0), a range containing only 0 that no card text produces.

Verification on the merged tree: cargo test -p phase-engine exit 0, cargo test
-p phase-ai exit 0, cargo clippy --workspace --all-targets --exclude probe-pin
exit 0, and the frontend suite including the seven-locale key-parity gate.
probe-pin is excluded because #7315 added it using Unix-only
ExitStatusExt::signal, which does not build on Windows; this branch does not
touch it.

+++ crates/engine/src/game/effects/reveal_chosen_numbers.rs
@@ -48,7 +48,7 @@ pub fn resolve(
     })
     .collect();
 
-    let mut numbers: Vec<(PlayerId, u8)> = Vec::new();
+    let mut numbers: Vec<(PlayerId, u32)> = Vec::new();
     for pid in candidates {
         if let Some(player) = state.players.iter_mut().find(|p| p.id == pid) {
             if let Some(value) = player.reveal_chosen_number() {
